Retreat on unwrapped ref front
It turned out that unwrapped references are more complicated
then I though. In particular use of applyDynamic doesn't make
it easy to redirect all methods calls with support for overloading
due to restrictions of current implementation.

There is a noteworthy hack by @xeno_by that might make it possible
with quite a bit of extra work though but it relies on private apis.

https://gist.github.com/xeno-by/8b64ca6a73775147941e