No More Mr Direct Region
It was originally introduced as a "solution" for applications
that need to allocate arbitrary-sized arrays that might not
necessarily fit into page pool model.

Unfortunately it was never a very performant solution due to
slowness of underlying malloc allocator plus an additional
overhead for bookkeeping to deallocate everything afterwards.

scala-offheap as a project strives to only provide high-performance
primitives and current DirectRegion implementation doesn't
fit this goal.